requirements.txt
boto3==1.9.123
botocore==1.12.123
certifi==2019.3.9
chardet==3.0.4
docutils==0.14
idna==2.8
jmespath==0.9.4
python-bittrex==0.3.0
python-dateutil==2.8.0
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.1
